Example of exposing internal func as pub. But scope is chaning to feature trait instead
